Fix execute permissions on build_release.sh
Git's bash shell on windows is build on MSYS2, which simulates unix-style fs
permissions. In particular, it guess executability based on some hueristics,
including whether the first line of a text file is a shebang (#!...) line. If it
is, then MSYS2 displays executability permissions, but they don't really exist
and are lost on git commit.

The workaround is to always update and commit permissions on a real *nix distro.