Make m_tags copyable
Adding a tg::tags::Tags member to a class makes that class uncopyable
due to the unique_ptr. Ryan thought this might be a workable approach.

This works but may be introducing other subtle issues. Might need to
revisit.

Ideally, I'd like to do a deep-copy that preserves the unique_ptr.